* push to branch and remove hosting for now as its not structured enough to useful

* fix minor issues

* More tweaking

* rename methods

* get rid of a MemoryMarshal and iterating using the reference as it makes it more readable

* Update events

* minor tweaks and safeguards

* add locking and begin support for adding questions and responses

* add question support, works for quiz questions, in theory
should work for other types but testing is required

* remove console write and add question support to HighLoadExample

* Remove external dependency and roll own user agent gen

* remove hosting for now (temporary)

* push some updates

* find optimal limits for large joins

* minor fixes

* remove framework dependent version and just use vector

* Add support for feedback & remove many uses of int in favor for uint

+++ Kahoot.NET/StateObject.cs
@@ -10,17 +10,17 @@ public class StateObject
     /// <summary>
     /// The size of the buffer to be used to store messages
     /// </summary>
-    public const int BufferSize = 1024;
+    public const int BufferSize = 1024; // 1 KiB
 
     /// <summary>
     /// The current message id count
     /// </summary>
-    public long id;
+    public ulong id;
 
     /// <summary>
     /// The number of acknowledgements
     /// </summary>
-    public long ack;
+    public ulong ack;
 
     /// <summary>
     /// The lag of the connection
@@ -57,6 +57,6 @@ public class StateObject
     public ExtWithTimesync<long> ExtWithTimesync => new()
     {
         Time = new() { L = l, O = o },
-        Acknowledged = ack
+        Acknowledged = (long)ack
     };
 }
